// src/app/core/models/system.model.ts
export interface ISystem {
  companyName: string;
  trademarkName: string;
  slogan: string;
  logo: string;
  favicon: string;
  iamge: string;
  copyright: string;
  status: boolean;
  CompanyAddress: string;
  transactionOffice: string;
  hotline: number;
  techniqueHotline: number;
  businessHotline: number;
  fax: number;
  email: string;
  taxCode: number;
  dunsCode: number;
  website: string;
  map: string;
  shotDescription: string;
  titleSeo: string;
  keySeo: string;
  imageSeo: string;
  descriptionSeo: string;
}